Defined stop of processing after finding a special marker in software.list.

+++ software-rebuilder/build-and-push-all.sh
@@ -10,6 +10,10 @@ note "Checking remote machine connection (shouldn't take more than a second).."
 ssh sofin@verknowsys.com "uname -a"
 
 for software in $(cat software.list); do
+    if [ "${software}" = "------" ]; then
+        note "Finished task."
+        exit 0
+    fi
     note "________________________________"
     note "Processing software: ${software}"
     s rm ${software}
